Remove unneeded gnuplot dependency
The graphicsmagick conda-forge package depends on gnuplot, which in
turn depends on a python interpreter.

AFAICS graphicsmagick doesn't depend / no-longer depends on gnuplot.

What I've found is that https://wiki.octave.org/GraphicsMagick lists
gnuplot as an optional dependency that

    ... and most users will have no need for them.

In the source code of GrpahicsMagick 1.3.45 I can still see that
gnuplot is part of configure.ac, but these lines are commented out.

Therefore, let's remove the gnuplot dependency from the package.

Fixes: #37